dj_shopping/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
Babel==2.12.1
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
cssselect2==0.7.0
Django==4.2
django-bootstrap5==23.1
django-phonenumber-field==7.1.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
django-weasyprint==2.2.0
django-wkhtmltopdf==3.4.0
fonttools==4.39.3
html5lib==1.1
idna==3.4
lxml==4.9.2
mysqlclient==2.1.1
oscrypto==1.3.0
pdfkit==1.0.0
phonenumbers==8.13.11
Pillow==9.5.0
pycparser==2.21
pydyf==0.6.0
pyHanko==0.18.1
pyhanko-certvalidator==0.22.0
pypdf==3.8.1
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.13
requests==2.30.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0b2
uritools==4.0.1
urllib3==2.0.2
weasyprint==58.1
webencodings==0.5.1
xhtml2pdf==0.2.11
zopfli==0.2.2
